stepper - unified and refactored component to be reused
