+++ web-jehlomat/src/organisms/organization/OrganizationLayout.tsx
@@ -1,15 +1,12 @@
-import { Box, useMediaQuery } from '@mui/material';
 import { PropsWithChildren } from 'react';
 import { Header } from '../../Components/Header/Header';
-import { secondaryColorVariant } from '../../utils/colors';
 import PenIcon from '../../assets/icons/pen.svg';
 import MessageIcon from '../../assets/icons/message.svg';
 import PenInactiveIcon from '../../assets/icons/penInactive.svg';
 import CheckInactiveIcon from '../../assets/icons/checkInactive.svg';
 import MessageInactiveIcon from '../../assets/icons/messageInactive.svg';
-import CheckSecondaryIcon from '../../assets/icons/checkSecondary.svg';
-import { SStep, SIcon, SSubTitle, SContainer } from './OrganizationLayout.styled';
-import { media } from '../../utils/media';
+import { StepperContainer, StepperCard } from 'Components/Stepper/Stepper';
+import { IconImage, SContainer } from './OrganizationLayout.styled';
 
 export enum RegistrationStep {
     CREATE,
@@ -25,37 +22,22 @@ interface Props {
 
 export function OrganizationLayout(props: PropsWithChildren<Props>) {
     const { step, title, children, backRoute } = props;
-    const isMobile = useMediaQuery(media.lte('mobile'));
 
     return (
         <>
             <Header backRoute={backRoute} mobileTitle={title || ''} />
 
-            {!isMobile && (
-                <Box minHeight={1} display="flex">
-                    <SStep initialStep={0 === step} active={0 <= step} flex={1} bgcolor={secondaryColorVariant('light')}>
-                        <SIcon active={0 <= step}>
-                            <img src={0 <= step ? PenIcon : PenInactiveIcon} alt="vytvoření účtu" />
-                        </SIcon>
-
-                        <SSubTitle>zadání údajů k vytvoření účtu</SSubTitle>
-                    </SStep>
-                    <SStep active={1 <= step} flex={1} bgcolor={secondaryColorVariant('regular')}>
-                        <SIcon active={1 <= step}>
-                            <img src={1 <= step ? MessageIcon : MessageInactiveIcon} alt="ověření e-mailové adresy" />
-                        </SIcon>
-
-                        <SSubTitle>ověření e-mailové adresy</SSubTitle>
-                    </SStep>
-                    <SStep active={2 <= step} flex={1} bgcolor={secondaryColorVariant('dark')}>
-                        <SIcon inverted active={2 <= step}>
-                            <img src={2 <= step ? CheckSecondaryIcon : CheckInactiveIcon} alt="úspěšné založení účtu" />
-                        </SIcon>
-
-                        <SSubTitle>úspěšné založení účtu</SSubTitle>
-                    </SStep>
-                </Box>
-            )}
+            <StepperContainer>
+                <StepperCard currentStep={step} order={0} title="zadání údajů k vytvoření účtu">
+                    <IconImage src={1 <= step ? PenIcon : PenInactiveIcon} alt="vytvoření účtu" />
+                </StepperCard>
+                <StepperCard currentStep={step} order={1} title="ověření e&#8209;mailové adresy">
+                    <IconImage src={2 <= step ? MessageIcon : MessageInactiveIcon} alt="ověření e-mailové adresy" />
+                </StepperCard>
+                <StepperCard currentStep={step} order={2} title="úspěšné založení účtu">
+                    <IconImage src={CheckInactiveIcon} alt="úspěšné založení účtu" />
+                </StepperCard>
+            </StepperContainer>
 
             <SContainer>{children}</SContainer>
         </>
